Move formUtils to utilities folder
