+++ src/content/questionnaire/sections/A.tsx
@@ -11,7 +11,7 @@ import TextInput from "../../../components/Questionnaire/TextInput";
 import AutocompleteInput from '../../../components/Questionnaire/AutocompleteInput';
 import institutionConfig from "../../../config/InstitutionConfig";
 import AddRemoveButton from '../../../components/Questionnaire/AddRemoveButton';
-import { filterNonNumeric, mapObjectWithKey } from '../utils';
+import { filterNonNumeric, mapObjectWithKey } from '../../../utils';
 import TransitionGroupWrapper from "../../../components/Questionnaire/TransitionGroupWrapper";
 import useFormMode from "./hooks/useFormMode";
