+++ src/content/questionnaire/sections/Review.tsx
@@ -4,7 +4,6 @@ import { parseForm } from "@jalik/form-parser";
 import { Divider, Grid, Stack, styled } from "@mui/material";
 import { useFormContext } from "../../../components/Contexts/FormContext";
 import { KeyedPlannedPublication, KeyedPublication } from "./B";
-import { mapObjectWithKey } from "../utils";
 import { KeyedContact } from "./A";
 import { KeyedFileTypeData } from "./D";
 import FormContainer from "../../../components/Questionnaire/FormContainer";
@@ -12,7 +11,7 @@ import ReviewSection from "../../../components/Questionnaire/ReviewSection";
 import ReviewDataListing from "../../../components/Questionnaire/ReviewDataListing";
 import ReviewDataListingProperty, { StyledValue } from "../../../components/Questionnaire/ReviewDataListingProperty";
 import ReviewFileTypeTable from "../../../components/Questionnaire/ReviewFileTypeTable";
-import { formatPhoneNumber } from "../../../utils";
+import { mapObjectWithKey, formatPhoneNumber } from "../../../utils";
 import useFormMode from "./hooks/useFormMode";
 import DataTypes from "../../../config/DataTypesConfig";
 import { StyledDescription, StyledTitle } from "../../../components/Questionnaire/SectionGroup";
